# contracts/control-plane-v1/control-plane.openapi.yaml
openapi: 3.1.0
info:
  title: AI Agent Desktop Local Control Plane API
  version: 1.0.0
  summary: Stable local management contract for the desktop GUI and trusted clients.
  description: |
    The read-only foundation, cc-connect Windows AMD64 isolated installation,
    product-owned non-secret configuration revisions, explicit lifecycle
    ownership, and managed process lifecycle are implemented. Real credential
    writes, Telegram binding, other component lifecycle, and GUI remain
    unsupported. Provider-specific structures remain behind Adapter boundaries.
  license:
    name: See repository LICENSE
    identifier: Apache-2.0
  x-contract-status: frozen
  x-implementation-status: cc-connect-managed-runtime-partially-verified
  x-event-types:
    - com.aiagentdesktop.component.state.changed.v1
    - com.aiagentdesktop.component.discovered.v1
    - com.aiagentdesktop.provider.state.changed.v1
    - com.aiagentdesktop.registry.changed.v1
    - com.aiagentdesktop.operation.started.v1
    - com.aiagentdesktop.operation.progress.v1
    - com.aiagentdesktop.operation.completed.v1
    - com.aiagentdesktop.operation.failed.v1
    - com.aiagentdesktop.operation.cancel_requested.v1
    - com.aiagentdesktop.scan.progress.v1
    - com.aiagentdesktop.plan.generated.v1
    - com.aiagentdesktop.diagnostic.created.v1
    - com.aiagentdesktop.task.state.changed.v1
    - com.aiagentdesktop.log.appended.v1
servers:
  - url: http://127.0.0.1:{port}/api/v1
...
    variables:
      port:
...
security:
  - localBearer: []
...
paths:
...
  "/components/{componentId}:install":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Components]
      operationId: installComponent
      summary: Install or repair one component after preflight.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
...
  "/components/{componentId}:uninstall":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Components]
      operationId: uninstallManagedComponentVersion
      summary: Uninstall one non-current product-managed cc-connect version.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UninstallComponentRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
  "/components/{componentId}:restore":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Components]
      operationId: restoreManagedComponentVersion
      summary: Restore the selected or previous product-managed cc-connect version.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComponentRestoreRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
...
  "/components/{componentId}/ownership-plans":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Configuration]
      operationId: createLifecycleOwnershipPlan
      summary: Create a separate explicit product lifecycle ownership handoff plan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                target_lifecycle_owner: { const: product }
                expires_in_seconds: { type: integer, minimum: 60, maximum: 3600, default: 900 }
      responses:
        "201":
          description: Ownership handoff plan.
          content:
            application/json:
              schema: { $ref: "./managed-runtime.schema.json#/$defs/OwnershipPlan" }
        default: { $ref: "#/components/responses/Problem" }
  "/components/{componentId}/ownership-plans/{planId}":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
      - name: planId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/Id" }
    get:
      tags: [Configuration]
      operationId: getLifecycleOwnershipPlan
      responses:
        "200":
          description: Ownership handoff plan.
          content:
            application/json:
              schema: { $ref: "./managed-runtime.schema.json#/$defs/OwnershipPlan" }
        default: { $ref: "#/components/responses/Problem" }
  "/components/{componentId}/ownership:confirm":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Configuration]
      operationId: confirmLifecycleOwnershipPlan
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [plan_id, plan_digest, current_management_owner, current_lifecycle_owner, confirmation]
              properties:
                plan_id: { type: string }
                plan_digest: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
                current_management_owner: { type: string, enum: [external, product, unmanaged, conflict, unknown] }
                current_lifecycle_owner: { type: string, enum: [external, product, none, conflict, unknown] }
                confirmation: { const: true }
      responses:
        "202": { $ref: "#/components/responses/OperationAccepted" }
        default: { $ref: "#/components/responses/Problem" }
  "/components/{componentId}/lifecycle":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    get:
      tags: [Components]
      operationId: getManagedLifecycleStatus
      summary: Probe and read product-managed process, owner, port, and health evidence.
      responses:
        "200":
          description: Managed runtime status. Deep health remains unsupported.
          content:
            application/json:
              schema: { $ref: "./managed-runtime.schema.json#/$defs/LifecycleRuntimeStatus" }
        default: { $ref: "#/components/responses/Problem" }
  "/components/{componentId}:reconcile":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Components]
      operationId: reconcileManagedLifecycle
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LifecycleRequest" }
      responses:
        "202": { $ref: "#/components/responses/OperationAccepted" }
        default: { $ref: "#/components/responses/Problem" }
  "/components/{componentId}/process-identity":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    get:
      tags: [Components]
      operationId: getManagedProcessIdentity
      responses:
        "200":
          description: Recorded identity and current verification result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  identity:
                    oneOf:
                      - { $ref: "./managed-runtime.schema.json#/$defs/ProcessIdentity" }
                      - { type: "null" }
                  verification: { type: [object, "null"] }
        default: { $ref: "#/components/responses/Problem" }
  "/components/{componentId}/port-ownership":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    get:
      tags: [Components]
      operationId: getManagedPortOwnership
      responses:
        "200":
          description: Exact TCP port-to-PID evidence, or null when not applicable.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "./managed-runtime.schema.json#/$defs/PortOwnershipEvidence" }
                  - { type: "null" }
        default: { $ref: "#/components/responses/Problem" }
  "/components/{componentId}:start":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Components]
      operationId: startComponent
      summary: Request the running target state.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
  "/components/{componentId}:stop":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Components]
      operationId: stopComponent
      summary: Request the stopped target state.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
  "/components/{componentId}:restart":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Components]
      operationId: restartComponent
      summary: Perform one serialized restart operation.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
...
  "/components/{componentId}:update":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Maintenance]
      operationId: updateComponent
      summary: Back up and update one component.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
  "/components/{componentId}:rollback":
    parameters:
      - $ref: "#/components/parameters/ComponentId"
    post:
      tags: [Maintenance]
      operationId: rollbackComponent
      summary: Roll back one component to a verified backup.
      x-stability: stable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollbackRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        default:
          $ref: "#/components/responses/Problem"
...
  "/operations/{operationId}":
    parameters:
      - $ref: "#/components/parameters/OperationId"
    get:
      tags: [Operations]
      operationId: getOperation
      summary: Read current progress and the mutually exclusive result or error.
      x-stability: stable
      responses:
        "200":
          description: Operation snapshot.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "./core-models.schema.json#/$defs/Operation"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    localBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-token
      description: At least 256 bits of entropy. Supplied only in the Authorization header.
  headers:
    ETag:
      description: Opaque resource revision for optimistic concurrency.
      schema:
        type: string
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Reusing a key with a different normalized request returns IDEMPOTENCY_KEY_REUSE.
      schema:
        type: string
        minLength: 16
        maxLength: 256
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/Id"
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Opaque expected revision. A stale value returns REVISION_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ComponentId:
      name: componentId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    OperationAccepted:
      description: Operation accepted. This does not prove that an external side effect has completed.
      headers:
        Location:
          description: Relative Operation resource URL.
          schema:
            type: string
        Retry-After:
          description: Suggested minimum polling interval in seconds when SSE is unavailable.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "./core-models.schema.json#/$defs/Operation"
    Problem:
      description: RFC 9457 compatible, user-facing and redacted problem details.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
...
    InstallRequest:
      type: object
      additionalProperties: false
      required: [version_policy, requested_version, source_ref, expected_digest, confirm, plan_id, plan_digest, confirmation]
      properties:
        version_policy:
          type: string
          enum: [exact, compatible, repair_current]
        requested_version:
          type: string
          maxLength: 128
        source_ref:
          $ref: "#/components/schemas/Id"
        expected_digest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        confirm:
          type: boolean
          const: true
        plan_id:
          $ref: "#/components/schemas/Id"
        plan_digest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        confirmation:
          type: boolean
          const: true
        deadline_at:
          type: string
          format: date-time
...
    UninstallComponentRequest:
      type: object
      additionalProperties: false
      required: [artifact_id, confirm]
      properties:
        artifact_id:
          $ref: "#/components/schemas/Id"
        confirm:
          type: boolean
          const: true
    ComponentRestoreRequest:
      type: object
      additionalProperties: false
      required: [confirm]
      properties:
        artifact_id:
          $ref: "#/components/schemas/Id"
        confirm:
          type: boolean
          const: true
    LifecycleRequest:
      type: object
      additionalProperties: false
      required: [configuration_revision, confirmation]
      properties:
        configuration_revision:
          type: integer
          minimum: 1
        confirmation:
          const: true
...
    UpdateRequest:
      type: object
      additionalProperties: false
      required: [target_version, backup_id, expected_digest, confirm]
      properties:
        target_version:
          type: string
          minLength: 1
          maxLength: 128
        backup_id:
          $ref: "#/components/schemas/Id"
        expected_digest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        confirm:
          type: boolean
          const: true
        deadline_at:
          type: string
          format: date-time
    RollbackRequest:
      type: object
      additionalProperties: false
      required: [backup_id, expected_revision, confirm]
      properties:
        backup_id:
          $ref: "#/components/schemas/Id"
        expected_revision:
          type: string
          minLength: 1
          maxLength: 128
        confirm:
          type: boolean
          const: true
        deadline_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, detail, user_message, retryable, recovery_actions, correlation_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
          maxLength: 256
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]{2,127}$"
        detail:
          type: string
          maxLength: 2048
        user_message:
          type: string
          minLength: 1
          maxLength: 2048
        retryable:
          type: boolean
        recovery_actions:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        diagnostic_id:
          $ref: "#/components/schemas/Id"
        operation_id:
          $ref: "#/components/schemas/Id"
        correlation_id:
          $ref: "#/components/schemas/Id"